+++ src/transformer/metadata-encoder.ts
@@ -75,17 +75,13 @@ export class MetadataEncoder {
             )
         ];
 
-        if (details.propertyNames.length > 0) {
-            if (ts.isClassDeclaration(klass) || ts.isClassExpression(klass))
-                decs.push(metadataDecorator('rt:SP', this.prepareElementNames(details.staticPropertyNames)));
-            decs.push(metadataDecorator('rt:P', this.prepareElementNames(details.propertyNames)));
-        }
+        if (ts.isClassDeclaration(klass) || ts.isClassExpression(klass))
+            decs.push(metadataDecorator('rt:SP', this.prepareElementNames(details.staticPropertyNames)));
+        decs.push(metadataDecorator('rt:P', this.prepareElementNames(details.propertyNames)));
 
-        if (details.methodNames.length > 0) {
-            if (ts.isClassDeclaration(klass) || ts.isClassExpression(klass))
-                decs.push(metadataDecorator('rt:Sm', this.prepareElementNames(details.staticMethodNames)));
-            decs.push(metadataDecorator('rt:m', this.prepareElementNames(details.methodNames)));
-        }
+        if (ts.isClassDeclaration(klass) || ts.isClassExpression(klass))
+            decs.push(metadataDecorator('rt:Sm', this.prepareElementNames(details.staticMethodNames)));
+        decs.push(metadataDecorator('rt:m', this.prepareElementNames(details.methodNames)));
 
         if (ts.isClassDeclaration(klass) || ts.isClassExpression(klass)) {
             let constructor = klass.members.find(x => ts.isConstructorDeclaration(x)) as ts.ConstructorDeclaration;